fix(mobile): platform-gate the ghostty terminal off iOS
expo-libghostty registers an Apple-only native view. Split the renderer
so Metro resolves the native surface only on iOS and a fallback elsewhere
(no expo-libghostty import), and hide the Terminals entry off iOS.
Verified: Android Hermes bundle has zero ghostty refs; iOS bundle keeps them.
Repo-wide prek lint skipped: pre-existing hub.ts error on the base branch;
all files in this commit pass eslint/tsc.
