+++ apps/mobile/src/app/host/[hostId]/index.tsx
@@ -5,7 +5,7 @@ import { EmptyState, ScreenScroll } from '@linkcode/ui/native';
 import { useLocalSearchParams, useRouter } from 'expo-router';
 import { Button, Card, Chip, Input, Label, ListGroup, Spinner, TextField } from 'heroui-native';
 import { useState } from 'react';
-import { RefreshControl, Text, View } from 'react-native';
+import { Platform, RefreshControl, View } from 'react-native';
 import { useTranslations } from 'use-intl';
 import { SessionStatusChip } from '../../../components/session-status-chip';
 
@@ -50,9 +50,16 @@ export default function SessionsScreen(): React.ReactNode {
       refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} />}
       headerRight={
         <View className="flex-row gap-1">
-          <Button variant="ghost" size="sm" onPress={() => router.push(`/host/${hostId}/terminal`)}>
-            <Button.Label>{t('terminals')}</Button.Label>
-          </Button>
+          {/* Terminals use the Apple-only ghostty native view; hide the entry off iOS. */}
+          {Platform.OS === 'ios' ? (
+            <Button
+              variant="ghost"
+              size="sm"
+              onPress={() => router.push(`/host/${hostId}/terminal`)}
+            >
+              <Button.Label>{t('terminals')}</Button.Label>
+            </Button>
+          ) : null}
           <Button variant="ghost" size="sm" onPress={() => router.push('/settings')}>
             <Button.Label>{t('settings')}</Button.Label>
           </Button>
